Update from ferret trunk as well, as fix in PyFerret version of tm_ftoc_readline.c, dealing with warnings raised by Ubuntu
git-svn-id: file:///home/users/tmap/svn/repos/ferret/branches/pyferret@14532 fdbf22ae-c210-0410-be80-ca943da6b8f8

+++ fer/ccr/EF_InternalUtil.c
@@ -3103,7 +3103,7 @@ int EF_ListTraverse_fprintf( char *data, char *curr )
    ExternalFunction *ef_ptr=(ExternalFunction *)curr; 
 
    fprintf(stderr, "path = \"%s\", name = \"%s\", id = %d, internals_ptr = %d\n",
-	   ef_ptr->path, ef_ptr->name, ef_ptr->id, ef_ptr->internals_ptr);
+	   ef_ptr->path, ef_ptr->name, ef_ptr->id, (long) (ef_ptr->internals_ptr));
 
    return TRUE;
 }
